Make tests/wide.rs exercise deep_safe, optionally with larger width.
Change tests_utils::sizes::great_width to return u64 so that larger values can
be given when desired, to demonstrate that the RecurQueue used by deep_safe has
inefficient memory usage when the width is larger, whereas wide_safe is
efficient.

+++ tests_utils/src/lib.rs
@@ -60,9 +60,9 @@ pub mod sizes
         get_env_var("MY_TEST_DEGENERATE_DEPTH", DEFAULT)
     }
 
-    pub fn great_width() -> u32
+    pub fn great_width() -> u64
     {
-        const DEFAULT: u32 = if cfg!(debug_assertions) { 100_000_000 } else { 1_000_000_000 };
+        const DEFAULT: u64 = if cfg!(debug_assertions) { 100_000_000 } else { 1_000_000_000 };
 
         get_env_var("MY_TEST_GREAT_WIDTH", DEFAULT)
     }
